@HippoBaro 2022-01-18
Upon request, I disabled the IO deduplication and reran the benchmarks. IO dedup made the numbers larger and less reliable.

### 10ms preemption:

Before: latency preemption alone
After: latency+throughput preemption

| Shares          | Before              | After                |
|-----------------|---------------------|----------------------|
| IO:100  CPU:100 | IOPS: 25k CPU: 157M | IOPS: 118k CPU: 111M |
| IO:100  CPU:10  | IOPS: 24k CPU: 159M | IOPS: 150k CPU: 92M  |

### 500us preemption:

Before: latency preemption alone
After: latency+throughput preemption

| Shares          | Before              | After               |
|-----------------|---------------------|---------------------|
| IO:100  CPU:100 | IOPS: 312k CPU: 15M | IOPS: 312k CPU: 15M |
| IO:100  CPU:10  | IOPS: 324k CPU: 9M  | IOPS: 336k CPU: 1M  |