@HippoBaro 2022-01-18
I forgot to mention that an essential knob of this new preemption is how many CQEs ready in the main ring trigger a preemption. The numbers I report above use a conservative value: the ring depth. Meaning that we only preempt if 128 CQEs are available for a ring with 128 SQE slots.

If we reduce this value, we get better numbers at the cost of more frequent preemptions.

Here are the same benchmark as above, but with this value set to 32 instead of 128:

### 10ms preemption:

Before: latency preemption alone
After: latency+throughput preemption (32 min CQEs available)

| Shares          | Before              | After                |
|-----------------|---------------------|----------------------|
| IO:100  CPU:100 | IOPS: 25k CPU: 157M | IOPS: 158k CPU: 90M |
| IO:100  CPU:10  | IOPS: 24k CPU: 159M | IOPS: 313k CPU: 7M  |

### 500us preemption:

Before: latency preemption alone
After: latency+throughput preemption (32 min CQEs available)

| Shares          | Before              | After               |
|-----------------|---------------------|---------------------|
| IO:100  CPU:100 | IOPS: 312k CPU: 15M | IOPS: 310k CPU: 13M |
| IO:100  CPU:10  | IOPS: 324k CPU: 9M  | IOPS: 324k CPU: 1M  |